+++ FRHED/FindUtil.h
@@ -27,6 +27,6 @@ Last change: 2013-02-24 by Jochen Neubeck
 #ifndef _FIND_UTIL_H_
 #define _FIND_UTIL_H_
 
-int findutils_FindBytes(BYTE* ps, int ls, BYTE* pb, int lb, int mode, bool case_sensitive);
+SSIZE_T findutils_FindBytes(BYTE* ps, SSIZE_T ls, BYTE* pb, SSIZE_T lb, int mode, bool case_sensitive);
 
 #endif // _FIND_UTIL_H_
