+++ FRHED/HexDump.h
@@ -40,9 +40,9 @@ public:
 			int byteSpace, int charSpace, int charset);
 	void SetOffsets(int minLen, int maxLen);
 	void SetArray(const SimpleArray<BYTE> *dataArray);
-	void CreateBuffer(unsigned size);
+	void CreateBuffer(size_t size);
 	char *GetBuffer() const;
-	void Write(int startInd, int endInd);
+	void Write(size_t startInd, size_t endInd);
 
 private:
 	SimpleArray<BYTE> *m_pData;
